--- src/libs/iOS/GameAnalytics.framework/GameAnalytics
@@ -1,1 +1,0 @@
-Versions/Current/GameAnalytics
