--- src/libs/iOS/GameAnalytics.framework/Headers
@@ -1,1 +1,0 @@
-Versions/Current/Headers
