--- src/libs/iOS/GameAnalytics.framework/Modules/module.modulemap
@@ -1,12 +1,0 @@
-framework module GameAnalytics {
-    umbrella header "GameAnalytics.h"
-
-    export *
-    module * { export * }
-
-    link "sqlite3"
-    link "z"
-
-    link framework "SystemConfiguration"
-    link framework "AdSupport"
-}
