--- src/libs/iOS/GameAnalytics.framework/Versions/Current
@@ -1,1 +1,0 @@
-A
