+++ src/libs/iOS/GameAnalytics.xcframework/ios-arm64/GameAnalytics.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module GameAnalytics {
+  umbrella header "GameAnalytics.h"
+  export *
+
+  module * { export * }
+}
