+++ src/libs/iOS/GameAnalytics.xcframework/ios-arm64_x86_64-simulator/GameAnalytics.framework/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module GameAnalytics {
+  umbrella header "GameAnalytics.h"
+  export *
+
+  module * { export * }
+}
