+++ src/libs/iOS/GameAnalytics.xcframework/tvos-arm64_x86_64-simulator/GameAnalyticsTVOS.framework/_CodeSignature/CodeResources
@@ -1,0 +1,132 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/GameAnalytics.h</key>
+		<data>
+		1wsToWD2A5H96dd5ZNCvI/YQbdo=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		YS1/gWwNEnmac3XJS2qgEwGSSMM=
+		</data>
+		<key>PrivacyInfo.xcprivacy</key>
+		<data>
+		TDHqRFjBXf3u63icf/o0aXE99V8=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/GameAnalytics.h</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			1wsToWD2A5H96dd5ZNCvI/YQbdo=
+			</data>
+			<key>hash2</key>
+			<data>
+			eMweH4OjfMXKcL+ANxH7TtgI5e43vTwdcPQmHkLHClc=
+			</data>
+		</dict>
+		<key>PrivacyInfo.xcprivacy</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			TDHqRFjBXf3u63icf/o0aXE99V8=
+			</data>
+			<key>hash2</key>
+			<data>
+			6VrI1/K2KCzAi1UIWJyxpaoAaRYkvm1xVDqH+AD4Al8=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
